+++ app/charts/pie/pie.tsx
@@ -131,6 +131,7 @@ export const Pie = () => {
             dy: 0,
             fontFamily,
             fontSize: labelFontSize,
+            strokeWidth: 8,
           }),
       });
     }