+++ app/docs/charts.stories.tsx
@@ -89,7 +89,7 @@ const ColumnsStory = {
       <InteractiveFiltersProvider chartConfigs={[chartConfig]}>
         <InteractiveFiltersChartProvider chartConfigKey={chartConfig.key}>
           <ColumnChart
-            limits={{ relatedDimension: undefined, limits: [] }}
+            limits={{ axisDimension: undefined, limits: [] }}
             observations={columnObservations}
             measures={columnMeasures}
             measuresById={keyBy(columnMeasures, (d: Measure) => d.id)}