+++ CHANGELOG.md
@@ -15,6 +15,8 @@ You can also check the
   - It's now possible to show segment value labels in column and bar charts
   - Axes now adapt dynamically to always show appropriate limits, even when they
     go beyond the data range
+  - Limits are now also rendered when the axis dimension is a single filter
+  - It's now possible to display limits in map charts
 - Fixes
   - Pie chart's `Measure` field is now correctly labeled (`Measure` instead of
     `Vertical Axis`)
@@ -24,6 +26,7 @@ You can also check the
     categorical dimensions
 - Styles
   - Optimized the custom map legends loading indicator appearance
+  - Line-based limits now take the full bar / column width
 
 # [5.4.0] - 2025-03-11